Reorganizing middleware
