+++ server/trpc.ts
@@ -64,74 +64,6 @@ const isAdmin = trpc.middleware(async ({ next, ctx }) => {
 });
 
 /* 
-This middleware is meant for routes that use a lectureId
-1. Look up the lecture using lectureId.
-2. Look up the courseMember using courseId and user email. Verify they are either
-a teacher or TA.
-3. If the courseMember is found the user has access.
-*/
-const isElevatedCourseMemberLecture = trpc.middleware(
-  async ({ next, ctx, rawInput }) => {
-    const email = ctx.session?.email;
-    const result = lectureInput.safeParse(rawInput);
-
-    if (!email)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'UNAUTHORIZED',
-          message: 'TRPC Middleware: User does not have a valid JWT'
-        })
-      );
-
-    if (!result.success)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'BAD_REQUEST',
-          message:
-            'TRPC Middleware: isElevatedCourseMemberLecture requires a valid lectureId'
-        })
-      );
-
-    const lecture = await prisma.lecture.findFirst({
-      where: {
-        id: result.data.lectureId
-      }
-    });
-
-    if (!lecture)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'TRPC Middleware: Lecture Not found'
-        })
-      );
-
-    //Find the first courseMember who is either a teacher or TA
-    const courseMember = await prisma.courseMember.findFirst({
-      where: {
-        courseId: lecture.courseId,
-        email: email,
-        OR: [
-          { role: zCourseRoles.enum.teacher },
-          { role: zCourseRoles.enum.ta }
-        ]
-      }
-    });
-
-    if (!courseMember)
-      throw generateTypedError(
-        new TRPCError({
-          code: 'UNAUTHORIZED',
-          message:
-            'TRPC Middleware: User either does not exist in course or does not have elevated priveleges'
-        })
-      );
-
-    return next();
-  }
-);
-
-/* 
 This middleware is meant for routes that use a courseId
 1. Look up the courseMember using courseId. Verify they are either a teacher or ta
 2. If the courseMember is found the user has access.
@@ -248,10 +180,6 @@ const isElevatedCourseMemberForCourseMember = trpc.middleware(
 export const router = trpc.router;
 export const publicProcedure = trpc.procedure;
 
-/* -------- Checks privileges using lectureId -------- */
-export const elevatedCourseMemberLectureProcedure = trpc.procedure.use(
-  isElevatedCourseMemberLecture
-);
 /* -------- Checks privileges using courseId -------- */
 export const elevatedCourseMemberCourseProcedure = trpc.procedure.use(
   isElevatedCourseMemberCourse
